+++ Editor/AteExternalCodeEditor.cs
@@ -1,3 +1,4 @@
+#if UNITY_EDITOR
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -151,3 +152,4 @@ namespace ADKOM.TextEditor
         }
     }
 }
+#endif