+++ Editor/TextDocument.cs
@@ -1,3 +1,4 @@
+#if UNITY_EDITOR
 using System;
 using System.IO;
 using System.Text;
@@ -164,3 +165,4 @@ namespace ADKOM.TextEditor
         };
     }
 }
+#endif