+++ Editor/UpdateChecker.cs
@@ -1,3 +1,4 @@
+#if UNITY_EDITOR
 using System;
 using UnityEditor;
 using UnityEditor.PackageManager;
@@ -208,3 +209,4 @@ namespace ADKOM.TextEditor
         }
     }
 }
+#endif